ahora uso spring security para verificar que el usuario que es autor o administrador pueda eliminar un post - solo faltaria subir imagen del post

+++ src/main/java/com/dani/roles/infrastructure/adapters/output/persistence/PostPersistenceAdapter.java
@@ -3,18 +3,24 @@ package com.dani.roles.infrastructure.adapters.output.persistence;
 import com.dani.roles.application.ports.output.PostPersistencePort;
 import com.dani.roles.domain.exception.UnauthorizedAccessException;
 import com.dani.roles.domain.model.Post;
+import com.dani.roles.domain.model.User;
 import com.dani.roles.infrastructure.adapters.output.persistence.entities.CategoryEntity;
 import com.dani.roles.infrastructure.adapters.output.persistence.entities.PostEntity;
 import com.dani.roles.infrastructure.adapters.output.persistence.entities.UserEntity;
 import com.dani.roles.infrastructure.adapters.output.persistence.mapper.PostPersistenceMapper;
+import com.dani.roles.infrastructure.adapters.output.persistence.mapper.UserPersistenceMapper;
 import com.dani.roles.infrastructure.adapters.output.persistence.repositories.CategoryRepository;
 import com.dani.roles.infrastructure.adapters.output.persistence.repositories.PostRepository;
 import com.dani.roles.infrastructure.adapters.output.persistence.repositories.UserRepository;
+import com.dani.roles.infrastructure.config.security.jwt.JwtUtil;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -24,30 +30,62 @@ public class PostPersistenceAdapter implements PostPersistencePort {
 
     private final PostRepository repository;
     private final PostPersistenceMapper mapper;
+    private final UserPersistenceMapper userMapper;
     private final CategoryRepository categoryRepository;
     private final UserRepository userRepository;
+    private final JwtUtil jwtUtil;
 
     @Override
     public Optional<Post> findById(Long id) {
-        return repository.findById(id).map(mapper::toPost);
+        return repository.findById(id).map(postEntity -> {
+            // Obtener el usuario asociado con el post
+            UserEntity userEntity = postEntity.getAuthor();
+            // Mapear UserEntity a User
+            User user = userMapper.toUser(userEntity);
+            // Mapear PostEntity a Post
+            Post post = mapper.toPost(postEntity);
+            // Establecer el usuario mapeado en el post
+            post.setUser(user);
+            return post;
+        });
     }
+
 
     @Override
     public List<Post> findAll() {
-        return mapper.toPostList(repository.findAll());
+        List<PostEntity> postEntities = repository.findAll();
+        List<Post> posts = new ArrayList<>();
+        for (PostEntity postEntity : postEntities) {
+            // Obtener el usuario asociado con cada publicación
+            UserEntity userEntity = postEntity.getAuthor();
+            // Mapear UserEntity a User
+            User user = userMapper.toUser(userEntity);
+            // Mapear PostEntity a Post
+            Post post = mapper.toPost(postEntity);
+            // Establecer el usuario mapeado en la publicación
+            post.setUser(user);
+            // Agregar la publicación a la lista
+            posts.add(post);
+        }
+        return posts;
     }
+
 
     @Override
     public Post save(Post post) {
+        // Obtener el usuario autenticado del contexto de seguridad
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (userDetails == null) {
+            throw new EntityNotFoundException("User details not found");
+        }
 
-        System.out.println("post user id PERSISTENCIA: " + post.getAuthorId());
-        System.out.println("post category id PERSISTENCIA: " + post.getCatIds());
+        String username = userDetails.getUsername();
 
         // Buscar si existe el usuario
-        UserEntity userEntity = userRepository.findById(post.getAuthorId())
-                .orElseThrow(() -> new EntityNotFoundException("User with ID " + post.getAuthorId() + " not found"));
-        // Asignar el usuario al post
-        post.setAuthorId(userEntity.getId());
+        UserEntity userEntity = userRepository.findByUsername(username);
+        if (userEntity == null) {
+            throw new EntityNotFoundException("User with username " + username + " not found");
+        }
 
         // Buscar si existe la categoría
         CategoryEntity categoryEntity = categoryRepository.findById(post.getCatIds())
@@ -70,7 +108,18 @@ public class PostPersistenceAdapter implements PostPersistencePort {
     public Post update(Long id, Post post) {
         //TODO -> Cuando implemente la seguridad, verificar que el usuario que intenta actualizar el post sea el mismo que lo creó
         // usando el contexto de seguridad de Spring Security, sin tener que asignar manualmente el id, sino del Bearer TOken
+        // Obtener el usuario autenticado del contexto de seguridad
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (userDetails == null) {
+            throw new EntityNotFoundException("User details not found");
+        }
 
+        String username = userDetails.getUsername();
+        // Buscar si existe el usuario
+        UserEntity userEntity = userRepository.findByUsername(username);
+        if (userEntity == null) {
+            throw new EntityNotFoundException("User with username " + username + " not found");
+        }
         // Verificar que el post exista
         if(repository.existsById(id)){
             // Obtener el post existente
@@ -80,7 +129,7 @@ public class PostPersistenceAdapter implements PostPersistencePort {
                     .orElseThrow(() -> new EntityNotFoundException("Category with ID " + post.getCatIds() + " not found"));
 
             // Verificar si el usuario que intenta actualizar el post es el mismo que lo creó
-            if(existingPost.getAuthor().getId().equals(post.getAuthorId())){
+            if(existingPost.getAuthor().getId().equals(userEntity.getId())){
                 // Actualizar los campos del post existente con los valores del post pasado como parámetro
                 existingPost.setTitle(post.getTitle());
                 existingPost.setDescription(post.getDescription());
@@ -100,6 +149,37 @@ public class PostPersistenceAdapter implements PostPersistencePort {
 
     @Override
     public void delete(Long id) {
-        repository.deleteById(id);
+        // Obtener los detalles del usuario actual
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (userDetails == null) {
+            throw new EntityNotFoundException("User details not found");
+        }
+
+        // Buscar el usuario en la base de datos
+        String username = userDetails.getUsername();
+        UserEntity userEntity = userRepository.findByUsername(username);
+        if (userEntity == null) {
+            throw new EntityNotFoundException("User with username " + username + " not found");
+        }
+
+        // Verificar si el usuario actual tiene el rol de administrador
+        boolean isAdmin = userEntity.getRoles().stream().anyMatch(role -> role.getName().equals("ADMIN"));
+
+        // Verificar si el post existe
+        if(repository.existsById(id)){
+            // Obtener el post existente
+            PostEntity existingPost = repository.findById(id)
+                    .orElseThrow(() -> new EntityNotFoundException("Post with ID " + id + " not found"));
+
+            // Verificar si el usuario actual es el autor del post o es administrador
+            if(existingPost.getAuthor().getId().equals(userEntity.getId()) || isAdmin){
+                repository.deleteById(id);
+            } else {
+                throw new UnauthorizedAccessException();
+            }
+        } else {
+            throw new EntityNotFoundException("Post with ID " + id + " not found");
+        }
     }
+
 }
